Refactor: 유저, 스토어 로그인 통합
