+++ src/main/java/com/stylelab/auth/presentation/response/SignInResponse.java
@@ -1,4 +1,4 @@
-package com.stylelab.store.presentation.response;
+package com.stylelab.auth.presentation.response;
 
 import com.stylelab.user.exception.UserError;
 import com.stylelab.user.exception.UserException;
@@ -6,7 +6,7 @@ import org.springframework.util.StringUtils;
 
 public record SignInResponse(String token) {
 
-    public static SignInResponse createResponse(String token) {
+    public static SignInResponse create(String token) {
         if (!StringUtils.hasText(token)) {
             throw new UserException(UserError.EMAIL_IS_REQUIRED);
         }
