--- src/main/java/com/stylelab/store/presentation/request/SignInRequest.java
@@ -1,25 +1,0 @@
-package com.stylelab.store.presentation.request;
-
-import com.stylelab.store.exception.StoreError;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Builder;
-
-@Builder
-public record SignInRequest(
-        @NotBlank(message = "STORE_STAFF_EMAIL_REQUIRE", payload = StoreError.class)
-        @Pattern(regexp = "^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}", message = "EMAIL_IS_NOT_IN_THE_CORRECT_FORMAT", payload = StoreError.class)
-        String email,
-        @NotBlank(message = "STORE_STAFF_PASSWORD_REQUIRE", payload = StoreError.class)
-        @Pattern(regexp = "^.*(?=^.{8,15}$)(?=.*\\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&+=]).*$", message = "PASSWORD_IS_NOT_IN_THE_CORRECT_FORMAT", payload = StoreError.class)
-        String password
-) {
-
-    @Override
-    public String toString() {
-        return "SignInRequest{" +
-                "email='" + email + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-}
