--- src/main/java/com/stylelab/user/presentation/request/SignInRequest.java
@@ -1,25 +1,0 @@
-package com.stylelab.user.presentation.request;
-
-import com.stylelab.user.exception.UserError;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Builder;
-
-@Builder
-public record SignInRequest(
-        @NotBlank(message = "EMAIL_IS_REQUIRED", payload = UserError.class)
-        @Pattern(regexp = "^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}", message = "EMAIL_IS_NOT_IN_THE_CORRECT_FORMAT", payload = UserError.class)
-        String email,
-        @NotBlank(message = "PASSWORD_IS_REQUIRED", payload = UserError.class)
-        @Pattern(regexp = "^.*(?=^.{8,15}$)(?=.*\\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&+=]).*$", message = "PASSWORD_IS_NOT_IN_THE_CORRECT_FORMAT", payload = UserError.class)
-        String password
-) {
-
-    @Override
-    public String toString() {
-        return "SignInRequest{" +
-                "email='" + email + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-}
